+++ graph/UndirectedGraph.cpp
@@ -16,7 +16,7 @@ void UndirectedGraph::loadGraph(vector<vector<int>> ajdList) {
 
     // check if the graph is a square matrix
     // check if the diagonal of the matrix is 0
-    for (int i = 0; i < ajdList.size(); i++) {
+    for (size_t i = 0; i < ajdList.size(); i++) {
         if (ajdList.size() != ajdList[i].size()) {
             throw invalid_argument("Invalid graph: The graph is not a square matrix.(" + std::to_string(i) + "th row has " + std::to_string(ajdList[i].size()) + " elements.)");
         }
@@ -28,8 +28,8 @@ void UndirectedGraph::loadGraph(vector<vector<int>> ajdList) {
     }
 
     // update the isWeighted and haveNegativeEdgeWeight fields if needed
-    for (int i = 0; i < ajdList.size(); i++) {
-        for (int j = i + 1; j < ajdList[i].size(); j++) {
+    for (size_t i = 0; i < ajdList.size(); i++) {
+        for (size_t j = i + 1; j < ajdList[i].size(); j++) {
             if (ajdList[i][j] != NO_EDGE) {
                 if (ajdList[i][j] != 1) {
                     this->isWeighted = true;
@@ -47,9 +47,9 @@ void UndirectedGraph::loadGraph(vector<vector<int>> ajdList) {
 
 void UndirectedGraph::printGraph() {
     int count_edges = 0;
-    for (int i = 0; i < ajdList.size(); i++) {
+    for (size_t i = 0; i < ajdList.size(); i++) {
         // only count the upper triangle of the matrix
-        for (int j = i + 1; j < ajdList[i].size(); j++) {
+        for (size_t j = i + 1; j < ajdList[i].size(); j++) {
             if (ajdList[i][j] != NO_EDGE) {
                 count_edges++;
             }
